added Auth

+++ src/tasks/tasks.service.ts
@@ -1,3 +1,4 @@
+import { User } from './../auth/user.entity';
 
 import { Injectable, NotFoundException, RequestTimeoutException } from '@nestjs/common';
 import { TasksController } from './tasks.controller';
@@ -17,8 +18,11 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) { }
 
-    getTasks(filterDto: GetTasksFilterDto):Promise<Task[]>{
-        return this.taskRepository.getTasks(filterDto)
+    getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User
+    ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user)
     }
     // getTaskWithStatus(filterDto: GetTasksFilterDto) {
     //     const { status, search } = filterDto;
@@ -38,16 +42,19 @@ export class TasksService {
     //     return this.tasks;
     // }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id)
+    async getTaskById(
+        id: number,
+        user: User
+    ): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id } })
         if (!found) {
             throw new NotFoundException(`Task with ID "${id}" not found`)
         }
         return found
     }
 
-    async createTask(createTaskDto: CreateTaskDto) {
-        return this.taskRepository.createTask(createTaskDto)
+    async createTask(createTaskDto: CreateTaskDto, user: User) {
+        return this.taskRepository.createTask(createTaskDto, user)
     }
 
     // createTask(createTaskDto: CreateTaskDto) {
@@ -62,16 +69,16 @@ export class TasksService {
     //     return task
     // }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id)
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id })
         console.log(result);
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID "${id}" not found`)
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id)
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user)
         task.status = status;
         await task.save()
